+++ wavesexchange_utils/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wavesexchange_repos"
+name = "wavesexchange_utils"
 version = "0.1.0"
 edition = "2021"
 authors = ["Artem Sidorenko <asidorenko@uniqsoft.ae>"]
@@ -7,14 +7,8 @@ authors = ["Artem Sidorenko <asidorenko@uniqsoft.ae>"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-envy = "0.4.2"
-serde = { version = "1.0.147", features = ["derive"] }
 thiserror = "1.0.38"
-tokio = { version = "1", default-features = false, features = ["sync", "rt-multi-thread"] }
 wavesexchange_log = { git = "https://github.com/waves-exchange/wavesexchange-rs", tag = "wavesexchange_log/0.5.1" }
 
 [dev-dependencies]
-tokio = { version = "1", features = ["macros"] }
-
-[features]
-config = []
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
