+++ wavesexchange_utils/src/circuit_breaker/error.rs
